Extract `ListsEditor` into a separate file
